Created the file layers
Now when the player presses enter after drawing their level they it is saved to a dictionary with list. All the empty space are filled with '-1' and the rest the csv value of their tile. These list now have to be transformed into csv files that can be saved on the device